We now have variable-length hipe_instructions although these are still transmitted and received as two args only.

+++ server/src/common.h
@@ -86,7 +86,7 @@ the completed instruction may be read from the output member variable.*/
 
 
 
-void decodeInstructionPreamble(const char* preamble, char* opcode, uint64_t* requestor, hipe_loc* location, uint32_t* arg1len, uint32_t* arg2len);
+void decodeInstructionPreamble(const char* preamble, char* opcode, uint64_t* requestor, hipe_loc* location, uint64_t* arg1len, uint64_t* arg2len);
